+++ tbot/tc/uboot/__init__.py
@@ -1,4 +1,3 @@
-from .build_info import BuildInfo
-from .build import build
+from .build import UBootBuilder, build, checkout
 
-__all__ = ("BuildInfo", "build")
+__all__ = ("UBootBuilder", "build", "checkout")